feat: many transfers Tx history columns
