+++ src/components/TransactionHistoryRows/TransactionGenericRow.tsx
@@ -1,37 +1,25 @@
 import { ArrowDownIcon, ArrowUpIcon } from '@chakra-ui/icons'
-import {
-  Box,
-  Button,
-  Circle,
-  Flex,
-  SimpleGrid,
-  Stack,
-  Tag,
-  Text as CText,
-  useColorModeValue,
-} from '@chakra-ui/react'
+import { Box, Button, Flex, SimpleGrid, Stack, Tag } from '@chakra-ui/react'
 import type { AssetId } from '@shapeshiftoss/caip'
 import { TradeType, TransferType } from '@shapeshiftoss/unchained-client'
 import { useMemo } from 'react'
 import { FaArrowRight, FaExchangeAlt, FaStickyNote, FaThumbsUp } from 'react-icons/fa'
 import { Amount } from 'components/Amount/Amount'
-import { AssetIcon } from 'components/AssetIcon'
 import { IconCircle } from 'components/IconCircle'
 import { Text } from 'components/Text'
 import { TransactionLink } from 'components/TransactionHistoryRows/TransactionLink'
 import { TransactionTime } from 'components/TransactionHistoryRows/TransactionTime'
 import type { Fee, Transfer } from 'hooks/useTxDetails/useTxDetails'
 import { Method } from 'hooks/useTxDetails/useTxDetails'
-import { bn, bnOrZero } from 'lib/bignumber/bignumber'
+import { bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
 import type { TxId } from 'state/slices/txHistorySlice/txHistorySlice'
 import { breakpoints } from 'theme/theme'
 
 import { ApproveIcon } from './components/ApproveIcon'
+import { AssetsTransfers } from './components/AssetsTransfers'
+import { AssetTransfer } from './components/AssetTransfer'
 import type { getTxMetadataWithAssetId } from './utils'
-
-const FALLBACK_PRECISION = 18
-const FALLBACK_SYMBOL = 'N/A'
 
 export const GetTxLayoutFormats = ({ parentWidth }: { parentWidth: number }) => {
   const isLargerThanSm = parentWidth > parseInt(breakpoints['sm'], 10)
@@ -99,91 +87,6 @@ type TransactionGenericRowProps = {
   parentWidth: number
 }
 
-type TransferStackProps = {
-  compactMode?: boolean
-  transfer: Transfer
-  index: number
-}
-
-type TransfersStackProps = {
-  compactMode?: boolean
-  transfers: Transfer[]
-  index: number
-}
-
-const TransferStack: React.FC<TransferStackProps> = ({ index, compactMode, transfer }) => {
-  return (
-    <Stack
-      alignItems='center'
-      key={index}
-      flex={1}
-      mt={{ base: 2, md: 0, xl: compactMode ? 2 : 0 }}
-      direction={index === 0 ? 'row' : 'row-reverse'}
-      textAlign={index === 0 ? 'left' : 'right'}
-    >
-      <AssetIcon
-        assetId={transfer.asset?.assetId}
-        boxSize={{ base: '24px', lg: compactMode ? '24px' : '40px' }}
-      />
-      <Box flex={1}>
-        <Amount.Crypto
-          color='inherit'
-          fontWeight='medium'
-          value={fromBaseUnit(transfer.value, transfer.asset?.precision ?? FALLBACK_PRECISION)}
-          symbol={transfer.asset?.symbol ?? FALLBACK_SYMBOL}
-          maximumFractionDigits={4}
-        />
-        {transfer.marketData.price && (
-          <Amount.Fiat
-            color='gray.500'
-            fontSize='sm'
-            lineHeight='1'
-            value={bnOrZero(
-              fromBaseUnit(transfer.value, transfer.asset?.precision ?? FALLBACK_PRECISION),
-            )
-              .times(transfer.marketData.price)
-              .toString()}
-          />
-        )}
-      </Box>
-    </Stack>
-  )
-}
-
-const TransfersStack: React.FC<TransfersStackProps> = ({ index, compactMode, transfers }) => {
-  const circleBgColor = useColorModeValue('white', 'whiteAlpha.100')
-  const circleColor = useColorModeValue('blue.100', 'blue.200')
-  const aggregatedFiatValue = transfers
-    .reduce((acc, transfer) => {
-      if (!transfer) return acc
-      return acc.plus(
-        bnOrZero(
-          fromBaseUnit(transfer.value, transfer.asset?.precision ?? FALLBACK_PRECISION),
-        ).times(transfer.marketData.price),
-      )
-    }, bn(0))
-    .toString()
-
-  return (
-    <Stack
-      alignItems='center'
-      key={index}
-      flex={1}
-      mt={{ base: 2, md: 0, xl: compactMode ? 2 : 0 }}
-      direction={index === 0 ? 'row' : 'row-reverse'}
-      textAlign={index === 0 ? 'left' : 'right'}
-    >
-      <Circle size={8} color={circleColor} borderWidth={2} bg={circleBgColor}>
-        <CText>{transfers.length}</CText>
-      </Circle>
-      <Box flex={1}>
-        <CText fontWeight='bold'>{`${transfers.length} Assets`}</CText>
-        <Amount.Fiat color='gray.500' fontSize='sm' lineHeight='1' value={aggregatedFiatValue} />
-      </Box>
-    </Stack>
-  )
-}
-
 export const TransactionGenericRow = ({
   type,
   title,
@@ -212,11 +115,11 @@ export const TransactionGenericRow = ({
 
       if (hasManyTypeTransfers) {
         return (
-          <TransfersStack index={index} compactMode={compactMode} transfers={typeTxTransfers} />
+          <AssetsTransfers index={index} compactMode={compactMode} transfers={typeTxTransfers} />
         )
       }
 
-      return <TransferStack index={index} compactMode={compactMode} transfer={transfer} />
+      return <AssetTransfer index={index} compactMode={compactMode} transfer={transfer} />
     })
   }, [compactMode, txTransfers, displayTransfers])
