+++ src/components/TransactionHistoryRows/constants.tsx
@@ -1,0 +1,2 @@
+export const FALLBACK_PRECISION = 18
+export const FALLBACK_SYMBOL = 'N/A'
